+++ src/Service/ReadingListManager.php
@@ -238,6 +238,7 @@ class ReadingListManager
                         'title' => $draft->title = (string)$tagValue,
                         'summary' => $draft->summary = (string)$tagValue,
                         'image' => $draft->image = (string)$tagValue,
+                        'author' => $draft->author = (string)$tagValue,
                         't' => $draft->tags[] = (string)$tagValue,
                         'a' => $draft->articles[] = (string)$tagValue,
                         default => null,